[feat] 백앤드 - SaaS hsbs-chat-bot UsageLog 관련 FastAPI 연동 로그 저장 Service 추가

+++ hsb-bo/src/main/java/com/hbs/hsbbo/admin/ai/usage/domain/entity/UsageLog.java
@@ -70,7 +70,7 @@ public class UsageLog extends AuditBase {
     private Integer totalTokens;
 
     @Column(name = "latency_ms")
-    private Integer latencyMs;           // 응답까지 걸린 시간(ms)
+    private Long latencyMs;           // 응답까지 걸린 시간(ms)
 
     // ============= 클라이언트 / 환경 정보 =============
